backend/requirements.txt
cachetools==5.3.2
certifi==2023.11.17
cffi==1.16.0
charset-normalizer==3.3.2
click==8.1.7
ffmpeg==1.4
google-api-core==2.15.0
google-auth==2.25.2
google-cloud-core==2.4.1
google-cloud-speech==2.23.0
google-cloud-translate==3.13.0
googleapis-common-protos==1.62.0
grpcio==1.60.0
grpcio-status==1.60.0
idna==3.6
joblib==1.3.2
nltk==3.8.1
numpy==1.26.2
proto-plus==1.23.0
protobuf==4.25.1
pyasn1==0.5.1
pyasn1-modules==0.3.0
PyAudio==0.2.14
pycparser==2.21
pydub==0.25.1
regex==2023.10.3
requests==2.31.0
rsa==4.9
simpleaudio==1.0.4
sounddevice==0.4.6
srt==3.5.3
tqdm==4.66.1
urllib3==2.1.0
webrtcvad==2.0.10
websockets==12.0
